Created a demo file to run simulations given a MatchingModel and some Policies.
Created a more general Policy called ThresholdsWithPriorities to be able
to test various policies changing what are the priorities and the
thresholds.

Created a new class EdgeData which is similar to NodesData but where
each value stored is related to an edge instead of a node. This is
useful to define the previous policy.

Created empty test files that should be filled in next commit with unit
test of the previous classes (policy and model).